requirements.txt
fastapi>=0.135.2
uvicorn>=0.42.0
pydantic>=2.12.5
langchain>=1.2.14
langchain-community>=0.4.1
langchain-core>=1.2.23
langchain-groq>=1.1.2
langchain-google-genai>=4.2.1
google-generativeai>=0.8.0
langchain-text-splitters>=1.1.1
chromadb>=1.5.5
PyMuPDF>=1.27.2.2
python-dotenv>=1.2.2
python-multipart>=0.0.22
numpy
pysqlite3-binary
